+++ src/engine/eval/mod.rs
@@ -10,6 +10,7 @@ pub use white_eval::WhiteEval;
 use crate::chess::game::Game;
 use crate::chess::piece::Piece;
 use crate::chess::square::Square;
+pub use crate::engine::eval::tapered_eval::PhasedEval;
 
 pub fn init() {
     piece_square_tables::init();
@@ -19,40 +20,30 @@ pub fn init() {
 pub struct IncrementalEvalFields {
     pub phase_value: i16,
 
-    pub midgame_pst_eval: WhiteEval,
-    pub endgame_pst_eval: WhiteEval,
+    pub piece_square_tables: PhasedEval,
 }
 
 impl IncrementalEvalFields {
     pub fn set_at(&mut self, sq: Square, piece: Piece) {
-        let phase_value_diff = tapered_eval::piece_phase_value_contribution(piece.kind);
-        self.phase_value += phase_value_diff;
-
-        let (mg, eg) = piece_square_tables::piece_contributions(sq, piece);
-        self.midgame_pst_eval += mg;
-        self.endgame_pst_eval += eg;
+        self.phase_value += tapered_eval::piece_phase_value_contribution(piece.kind);
+        self.piece_square_tables += piece_square_tables::piece_contributions(sq, piece);
     }
 
     pub fn remove_at(&mut self, sq: Square, piece: Piece) {
-        let phase_value_diff = tapered_eval::piece_phase_value_contribution(piece.kind);
-        self.phase_value -= phase_value_diff;
-
-        let (mg, eg) = piece_square_tables::piece_contributions(sq, piece);
-        self.midgame_pst_eval -= mg;
-        self.endgame_pst_eval -= eg;
+        self.phase_value -= tapered_eval::piece_phase_value_contribution(piece.kind);
+        self.piece_square_tables -= piece_square_tables::piece_contributions(sq, piece);
     }
 }
 
 impl IncrementalEvalFields {
     pub fn init(board: &Board) -> Self {
         let phase_value = tapered_eval::phase_value(board);
-        let (midgame_pst_eval, endgame_pst_eval) = piece_square_tables::phase_evals(board);
+        let piece_square_tables = piece_square_tables::eval(board);
 
         Self {
             phase_value,
 
-            midgame_pst_eval,
-            endgame_pst_eval,
+            piece_square_tables,
         }
     }
 }
@@ -63,14 +54,9 @@ pub fn eval(game: &Game) -> Eval {
 }
 
 pub fn absolute_eval(game: &Game) -> WhiteEval {
-    let midgame_eval = game.incremental_eval.midgame_pst_eval;
-    let endgame_eval = game.incremental_eval.endgame_pst_eval;
+    let eval = game.incremental_eval.piece_square_tables;
 
-    tapered_eval::taper(
-        game.incremental_eval.phase_value,
-        midgame_eval,
-        endgame_eval,
-    )
+    tapered_eval::taper(game.incremental_eval.phase_value, eval)
 }
 
 #[derive(Debug)]
@@ -84,16 +70,15 @@ pub struct EvalComponents {
 
 pub fn eval_components(game: &Game) -> EvalComponents {
     let eval = absolute_eval(game);
-
-    let (midgame_pst, endgame_pst) = piece_square_tables::phase_evals(&game.board);
     let phase_value = tapered_eval::phase_value(&game.board);
 
-    let pst_eval = tapered_eval::taper(phase_value, midgame_pst, endgame_pst);
+    let piece_square_tables = piece_square_tables::eval(&game.board);
+    let pst_eval = tapered_eval::taper(phase_value, piece_square_tables);
 
     EvalComponents {
         eval,
-        piece_square_midgame: midgame_pst,
-        piece_square_endgame: endgame_pst,
+        piece_square_midgame: piece_square_tables.midgame(),
+        piece_square_endgame: piece_square_tables.endgame(),
         phase_value,
         piece_square_tables: pst_eval,
     }