+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Use a dedicated `MoveList` struct instead of `Vec<Move>`
 * Store castle rights as an array indexed by player
+* Bundle midgame and endgame evals into a `PhasedEval` struct
 
 ### Misc